sched: WakeMigration: Add cpu_saturated and saturated cpu accounting
INTRODUCTION
============

This patch introduces a new member in sg_lb_stats,
nr_saturated, which is an indicator of the number of
saturated CPUs in the sched_group.

A saturated CPU is defined as a CPU which is being overutilized
by a single task. The criteria for saturation is

	utilization_load_avg		capacity_margin
       ----------------------     <   ------------------
    capacity_of(cpu) * nr_running	    1024

The goal of nr_saturated is to identify if a group has CPUs
that are running tasks which could benefit from being moved to
a group with higher capacity CPUs.

TARGET PROBLEM
==============

Since the tasks that tend to saturate CPUs can potentially
be CPU hogging and may not undergo wakeups to eventually land up
on the correct/fitting CPU, such an intervention is necessary even
if the group has a single running task.

Consider a scenario when a task changes it's nature from being
I/O bound, i.e waits for some resources to being CPU bound. In
such a scenario, where this is the only task running on the CPU
the find_busiest_group returns NULL which does not trigger the
load balancer and thus does not migrate to a higher capacity CPU

In order to identify such scenarios, where the CPU is being saturated
by a single task, cpu_saturated is being introduced.

Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>