WIP: SchedTune: add TC2 energy model normalization constants
The energy_diff() function compute the "absolute" energy variation, estimated,
by using the energy model data, to switch between two different scheduling
decisions. Being this an "absolute" value it is not suitable to define
a scheduling acceptance strategy which is based on the evaluation of
relative Power-vs-Performances variations.

To support such a strategy we need a "relative" metric which describes
how much the energy is impacted by a certain scheduling decision.
This impact must be defined in percentage terms, with respect to the maximum
(CPUs) energy consumption which could be generated by the specific target.

This patch adds the definition of the min/max energy (CPUs) consumptions
for the specific TC2 platform. These data could be obtained by computation
from the energy model and sched domain data however, for the time being,
we hard-coded them in the SchedTune module.

The patch adds also the couple of constants "nrg_shift" and "nrg_mult" which
could be used to implement a fast integer division (by constant) to get
the normalized energy consumption given the max/min energy consumption of the
platform. Such a fast division could be computed using the formula:

	energy / energy_max ~= (energy * nrg_mult) >> nrg_shift

TODO: add an initialization function which compute normalization constants
      from Energy Model (EM) and Sched Domains (SD) data.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>