WIP: SchedTune: add support to compute normalized energy_diff
The energy_diff() function compute the "absolute" energy variation, estimated,
by using the energy model data, to switch between two different scheduling
decisions. Being this an "absolute" value it is not suitable to define
a scheduling acceptance strategy which is based on the evaluation of
relative Power-vs-Performances variations.

To support such a strategy we need a "relative" metric which describes
how much the energy is impacted by a certain scheduling decision.
This impact must be defined in percentage terms, with respect to the maximum
(CPUs) energy consumption which could be generated by the specific target.

This patch adds a new function which exploits target specific normalization
constants to compute a "relative" energy variation starting from the
"absolute" value predicted by the energy model data.

The "relative" energy variation is a new metrics which is:
- defined in the range [0..1024]
- is 0 when the estimated energy_diff is