WIP: EXPERIMENTAL: sched: add energy_diff_evaluate to wakeup path
Once the SchedTune support is enabled and the CPU bandwidth demand of a
task is boosted, we could expect an increased energy consumption which is
balanced by a corresponding increase of the task performances.
However, the current implementation of the energy_diff() function accepts
all and _only_ the schedule candidates which results into a reduced
expected system energy. This will work against the boosting strategy.

This patch links the energy_diff() function with the "energy payoff" engine
provided by SchedTune.
The energy variation computed by the energy_diff() function is now filtered
using the SchedTune support to evaluated the energy payoff for a boosted
task.
All and only the schedule candidate which corresponds to a positive
energy_payoff are now going to be reported as "more energy efficient"
by the energy_diff() function.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>