FIXUP: sched: avoid computations when moving small tasks
In general it seems to make sense to avoid the computation of the
energy_diff() function when moving really small tasks.
Perhaps, these tasks should just be moved to a LITTLE CPU, maybe the first
one, which could be easily determined or defined at boot time as the CPU
with the highest energy efficiency.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>