FIXUP: sched/cpufreq_sched_cfs: fix task_tick_fair to consider RQ load
One of the goals of SchedDVFS is to switch to the maximum OPP once the
actual CPU usage is approaching an upper bound margin relative to the
current OPP capacity.

We check the distance from this margin considering the current task
utilization. This solution is expected to work properly only if we have a
single task on the RQ. However, when we have multiple (perhaps big) tasks
concurrently active on a CPU, since they are multiplexed by the scheduler,
they could not build up an active utilization which is big enough to
trigger an OPP switch.

One solution could be that to consider the CPU usage, which is the sum of
active and blocked utilization on a CPU.

NOTE: A better solution should probably be to consider just the CPU
utilization, i.e. disregarding the blocked utilisation.
Indeed, only concurrently active tasks are preempted and not allowed to
build up a utilization to trigger the OPP change, but still their
cumulative utilization is tracked by the RQ signal.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>