WIP: sched/tune: compute and keep track of per CPU boost value
When per task boosting is enabled, we could have multiple RUNNABLE tasks
which are concurrently scheduled on the same CPU but each one with
a different boost value. For example, we could have a scenarios like
that:

  Task   SchedTune CGroup   Boost Value
    T1               root            0
    T2       low-priority           10
    T3        interactive           90

In these conditions we expect a CPU to be configured according to a
proper "aggregation" of the required boost values for all the tasks
currently scheduled by that CPU.

A suitable aggregation function is the one which tracks the MAX boost
value for all the tasks RUNNABLE on a CPU. This approach allows to
always satisfy the most boost demanding task while at the same time:
 a) boosting all the concurrently scheduled tasks thus reducing
    potential co-scheduling side-effects on demanding tasks
 b) reduce the frequency switching required to SchedDVFS thus being more
    friendly to architectures with slow frequency switching times

Every time a task enter/exit the RQ of a CPU the max boost value should
be potentially updated considering all the boost groups currently
"affecting" that CPU, i.e. because they have at least one RUNNABLE task
currently allocated on that CPU.

This patch introduces the required support to keep track of the boost
groups currently affecting each CPU. The provided implementation is
quite simple. Indeed, thanks to the limited number of boost groups which
could be allocated on a system, a small and memory efficient per-cpu
array of boost groups values (cpu_boost_groups) is used which is
updated, for each CPU entry, by schedtune_boostgroup_update() but just
when a schedtune CGroup boost value is updated.
However, this is expected to be a not frequent operation, perhaps done
just one time at system boot time.

Change-Id: I4065438aa3b24731508dbb4ce4ce0278b306dacc
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>