sched/tune: add sysctl interface to define a boost value
The current scheduler implementation does not allow "to boost" tasks
performance by running them at higher OPPcompared to the minimum
required to meet their workload demands.

To support tasks performance boosting the scheduler should to provide a
"knob" which allows to tune how much the system is going to be optimised
for energy efficiency vs performance.

This patch is the first of a series which provides a simple interface to
define a tuning knob. One system-wide "boost" tunable is exposed via:
  /proc/sys/kernel/sched_cfs_boost
which can be configured in the range [0..100], to define a percentage
where:
  - 0%   boost requires to operate in "standard" EAS mode by scheduling
         tasks at the minimum capacities required by the workload demand
  - 100% boost requires to push at maximum the task performances,
         "regardless" of the incurred energy consumption

A boost value in between these two boundaries is used to bias the
power/performance trade-off, the higher the boost value the more the
scheduler is biased toward performance boosting instead of energy
efficiency.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>