WIP: schedtune: add sysctl interface to define a boost value
The energy-aware scheduler extension has been designed to exploit an
energy model to support an energy efficient allocation of tasks on available
CPUs. The main goal of the current implementation is to schedule
tasks in such a way to minimise the expected system energy while still
meeting the requirements of tasks in terms of computational power.
Thus, the current implementation does not allow "to boost" tasks performances,
for example by running them at an higher OPP (or a more capable CPU), even
if that could require a "reasonable" increase in energy consumption.

To support tasks performance boosting, while still operating in energy-aware
mode, the scheduler should to provide a "knob" which allows to tune how
much the system is going to be optimised for energy efficiency vs performances.

This patch is the first of a series which provides a simple sysctl based
interface to define an EAS tuning knobs.
For the time being, just one system-wide "boost" tunable is exposed via:
  /proc/sys/kernel/sched_cfs_boost
which can be configured in the range [0..100], to define a percentage
where:
  - 0%   boost requires to operate in "standard" EAS mode by scheduling
         tasks at the minimum capacities required by the workload demand
  - 100% boost requires to push at maximum the task performances,
         "regardless" of the incurred energy consumption

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>