WIP: schedtune: add function to convert the boost value into a "margin" signal
The basic idea of the boost knob is to "artificially inflate" some signals
to make a Task/RQ appears more demanding than what actually it is.
Independently from the specific signal, a consistent and possibly simple
semantic for the concept of "signal boosting" should be defined.
Such a semantic must defined:
1. how we translate the boost percentage into a "margin" value to be added
   to the original signal to be inflated
2. what is the meaning of a boost value from a user-space perspective

This patch provides the implementation of a possible boost semantic,
named Signal Proportional Compensation (SPC), where the boost percentage
(BP) is used to compute a margin (M) which is proportional to the
complement of the original signal (OS):
  M = BP * (1024 - OS)
The computed margin then added to the OS to obtain the Boosted Signal (BS)
  BS = OS + M

The proposed boost semantic has these main features:
- each signals gets a boost which is proportional to its delta with respect
  to the maximum available capacity in the system
- a 100% boosting has a clear understanding from a user-space perspective,
  since it means simply to run (possibly) "all" tasks at the max OPP
- each boosting value means to improve the tasks performance by a quantity
  which is proportional to the maximum achievable performance on that
  system.
Thus this semantics is somehow forcing a behaviour which is:

  50% boosting means to run at half-way between the current and the
    maximum performance which a task could achieve on that system

This patch provides the code to implement a fast integer division to convert
a boost percentage (BP) value into a margin (M).

NOTE: this code is suitable for all signals operating in range
      [0..SCHED_LOAD_SCALE]

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>