sched/fair: add boosted task utilization
The task utilization signal, which is derived from PELT signals and properly
scaled to be architecture and frequency invariat, is used by EAS as an
estimation of the task requirements in terms of CPU bandwidth.
This signal affects both the CPU selection as well as, when SchedDVFS is
in use, the selection of the current operating points (OPP) for the CPU.
A convenient way to bias these decisions, which is also little intrusive,
is to boost the task utilization signal each time must be used to support
them.

This patch introduces the new function:
  boosted_task_utilization(task)
which returns a boosted value for the utilization of the specified task.
The margin added to the original utilization is:
  1. computed by the boosting strategy introduced by a previous patch
  2. proportional to the system-wide boost value defined by the sysctl
     interface, this also introduced by a previous patch

The boosted signal is used by EAS
  a. transparently, via its integration into the task_fits() function
  b. explicitly, in the energy-aware wakeup path

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>