sched/fair: keep track of energy/capacity variations
The current EAS implementation does not allow "to boost" tasks performances,
for example by running them at an higher OPP (or a more capable CPU), even
if that could require a "reasonable" increase in energy consumption.
To defined how much reasonable is an energy increase with respect to a
required boost value, it is required to define and compute a trade-off
between the expected energy and performance variations.
However, the current EAS implementation considers only energy variations
while completely disregard the impact on performance for the selection of
a certain schedule candidate.

This patch extends the eenv energy environment to keep track of both energy
and performance deltas which are implied by the activation of a schedule
candidate.
The performance variation is estimated considering the different capacities
of the CPUs in which the task could be scheduled. The idea is that while
running on a CPU with higher capacity (e.g. higher operating point) the
task could (potentially) complete faster and thus get better performance.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>