sched/tune: add support to compute normalized energy
The current EAS implementation not only considers only energy variations,
while completely disregard the impact on performance for the selection of
a certain schedule candidate, but it also compute "absolute" energy variations.
In order to properly define a trade-off strategy between increased energy
consumption and performances benefits it is required to compare energy
variations with performance variations.
Thus, both these metrics must be expressed in comparable units.
While the performance variations are expressed in terms of capacity deltas,
which are expressed in the range [0..SCHED_LOAD_SCALE], the same scale is
not used for energy variations thus

This patch introduces a new the support function:
  schedtune_normalize_energy(energy_diff)
which returns a normalized value in the same range of capacity variations,
i.e. [0..SCHED_LOAD_SCALE].

NOTE: energy normalization requires some data from the Energy Model (EM) of
      the specific target. Since these values are expected to be provided
      at boot time along with the EM itself, for the time being this patch
      hard-code the values of the same ARM TC2 EM provided by the previous
      EAS patches.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>