WIP: sched/tune: add performance-energy space filtering function
Boosting the utilisation of a task could imply the selection of a CPU/OPP
and cluster which has an higher energy impact. In that conditions the
energy_diff() function will discard the scheduling proposal under evaluation
thus impacting the possibility to boost the performances of a task.
A trade-off between increased energy consumption and performances benefits
should be evaluated.

This patch introduces the required support to evaluate the "energy payoff"
corresponding to a certain variation on expected task performances.
A proposed schedule candidate is described in term of expected variation
of energy and performance. These values are used to verify in which region
of the Performance-Energy Space the solution becomes and thus returns
an energy_payoff metrics which is positive if
- the increased energy is compensated by a configured performance gain
  i.e. either in Optimal (O) or Performance Boost (B) region
- the decreased energy is compliant with a configured performance constraint
  i.e. in Performance Constraint (C) region

Otherwise the returned energy_payoff will be negative, i.e. not in the above
cases or in the Suboptimal (S) region.

The classification of a schedule decision in one of the four regions is
defined also by a couple of thresholds in the Performance-Energy variation
Space, which allows to identify two main optimization subregions:
- Performance Boost (B) region
  top right quadrant: both nrg and cap expected variations are positive
- Performance Constraining (C) region
  bottom-left quadrant: both nrg and cap expected variations are negative

This implementation binds both the thresholds values to the single value
of the boost margin, thus allowing to use a single boost knob to configure
all the thresholds.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

Conflicts:
	kernel/sched/tune.c