sched/{fair,tune}: add initial support for CGroups based boosting
To support tasks performance boosting, while still operating in
energy-aware mode, the previous patches introduces a system-wide "knob"
which allows to tune how much the system is going to be optimized for
energy efficiency vs performances.
The usage of a single knob has the advantage to be a simple solution,
both from the implementation and the usage standpoint.
However, on a real system it is in general difficult to identify a
single value for the knob which could fits multiple different tasks.
For example, some kernel threads and/or user-space background services
should be better managed always in an energy efficient way while still
being able to boost the performance of specific workloads.

In order to improve the flexibility of the task boosting mechanism this
patch is the first of a small series which extends the previous
implementation to introduce a "per task group" support.
The new "schedtune" CGroups controller is added by this patch which
allows to configure different boost value for different groups of tasks.
To keep the implementation simple while still being effective for a
boosting strategy, the new controller:
  1. allows only a two layer hierarchy
  2. supports only a small number of boost groups

A two layer hierarchy allows to place each task either:
  a) in the root control groups
     thus being subject to a system-wide boosting value
  b) in a child of the root group
     thus being subject to the specific boost value defined by that
     "boost group"
This decision is based on the observation that in general it is
difficult to define a clear user-space semantic for nested groups.

The limited number of "boost groups" supported is mainly motivated by
the observation that in a real system it could be useful to have only
few classes of tasks which deserve different treatment.
For example, background vs foreground or interactive vs low-priority.
As an additional benefit, a limited number of boost groups allows also
to have a more simple implementation especially for the code required to
compute the boost value for CPUs which have runnable tasks belonging to
different boost groups.

This first patch introduces just the basic CGroups support as well as an
updated version of the schedtune_accept_deltas() function which filter
the energy-diff value considering the specific boost value assigned to
the task being evaluated.
Since CPUs boosting deserves a new set of updates, with respect to the
system-wide approach, it is going to be introduced by a following
patch.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>