sched/fair: jump to max OPP when crossing UP threshold
To make free room for a task that is building its utilization, to harm
its performance the least and because we won't know the "real" utilization
of the task if it continues saturating the current OPP, jump to max OPP
as soon as get_cpu_usage() (which contains usage of all tasks on the rq)
crosses the UP threshold. This threshold is built using the same margin
(capacity_margin) used to tell if a cpu is overutilized.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>