sched/fair: add triggers for OPP change requests
Each time a task is {en,de}queued we might need to adapt the current
frequency to the new usage. Add triggers on {en,de}queue_task_fair()
for this purpose. Only trigger a freq request if we are effectively
waking up or going to sleep. Filter out load balancing related calls
to reduce the number of triggers.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>