sched/{core,fair}: trigger OPP change request on fork()
Patch "sched/fair: add triggers for OPP change requests" introduced OPP
change triggers for enqueue_task_fair(), but the trigger was operating
only for wakeups. Fact is that it makes sense to consider wakeup_new also
(i.e., fork()), as we don't know anything about a newly created task and
thus we most certainly want to jump to max OPP to not harm performance
too much.

However, it is not currently possible (or at least it wasn't evident to
me how to do so :/) to tell new wakeups to other (non wakeup) operations.

This patch introduces an additional flag in sched.h that is only set
at fork() time and it is then consumed in enqueue_task_fair() for our
purpose.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>