CHROMIUM: arm64: dts: mt8173-oak-rev1: Disable highest 2 OPP
The MT8173 chips on some oak-rev1 boards are engineering samples that do
not support the two highest OPPs for the big core.

Since there is no way to tell at runtime whether the particular MT8173
supports these OPPs or not, all we can do is disable them for all
oak-rev1 boards.

Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@mediatek.com>

BUG=chrome-os-partner:39501
TEST=build and boot oak-rev1, -rev2 and test each OPP of two clusters

Change-Id: I4b079f1a0df02cd1005be85fc0ff00b01532b4e1