sched/{fair,cpufreq_sched}: use static_key_deferred
Mike commented:

"<paranoia>

These controls are exposed to userspace via cpufreq sysfs knobs. Should
we use a struct static_key_deferred and static_key_slow_dec_deferred()
instead? This helps avoid a possible attack vector for slowing down the
system.

</paranoia>

I don't really know what a sane default rate limit would be in that case
though."

Which makes probably sense. So, switch to use that. We choose a rate
limit of HZ, as it seems to be the common default.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>