sched/fix: 16_46 Initialize sd energy function to NULL by default
Initialize sched_domain_topology_level::sched_domain_energy_f to NULL by
default for all architectures.
Do this for sched_domain_topology_level::sched_domain_flags_f as well.

This has to be done for ARM and ARM64 as well and later in the patch
series we (18/46 arm: topology: Define TC2 energy and provide it to the
scheduler) when we introduce the energy model, the sd energy function has
to be initialized appropriately.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>