sched/fix: 26_46 Check if cpu_map is not empty in build_sched_domains
In case there is a cpuset cgroup with 0 cpus in it and load balancing will
be disabled for the root cpuset cgroup, build_sched_domains() is called
with an empty cpu_map which leads to a crash when calling
cpu_rq(cpumask_first(cpu_map).

Rather than guarding this by 'if(!cpumask_empty(cpu_map))' or '(if
cpumask_first(cpu_map) < NR_CPUS)' initialize the local variable rq to
NULL and use 'if (rq)'. If rq != NULL cpu_map must not be empty.

Credit to Jury for discovering this perfidious bug!

How-to recreate this bug:

 cgcreate -g cpuset:a

 echo 0 > /sys/fs/cgroup/cpuset/cpuset.sched_load_balance

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>