sched/fix: 16/46 Let sched_domain_energy_f function return const pointer
Let the sched_domain_energy_f function return a constant 'struct
sched_group_energy' pointer to a constant 'struct sched_group_energy'.

This is an indirect suggestion by PeterZ during the RFCv5 review.

So the topolgy layer or the scheduler can't change the content of the
sched group energy data but it can't change the pointer to it embedded in
struct sched_group either.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>