sched/fair: Scale load/util contribution rather than time
When updating load/util tracking the time delta might be very small in
many cases, scaling it futher down with frequency and cpu invariance
scaling might cause us to loose precision. Instead of scaling time we
can scale the weight of the task for load and the capacity for
utilization. Both weight and capacity should be significantly bigger in
most cases, but low priority tasks might still suffer.

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>