FIXUP: sched/cpufreq_sched: Clear __sched_energy_freq only when no policy is using sched governor
The current implementation of SchedDVFS is turned off completely if a
cpufreq policy is completely offline. This is because the way the
{set|clear}_sched_energy_freq() functions are implemented only allows
the __sched_energy_freq static_key to be either 1 or 0, while there
are two possible cpufreq policies (one for each cluster) on Oak. The
__sched_energy_freq static_key acts like a switch for the SchedDVFS
hook in the CFS scheduler. So whenever a cluster is turned off, which
sets the corresponding cpufreq policy offline, will result in the
__sched_energy_freq being set to 0.

Signed-off-by: Ricky Liang <jcliang@chromium.org>
[steve.muckle@linaro.org: rebase atop change to name of static key]
Signed-off-by: Steve Muckle <steve.muckle@linaro.org>