MALI: drm/cma-helper: Add multi buffer support for cma fbdev
This patch add a config to support to create multi buffer for cma fbdev.
Such as double buffer and triple buffer.

Cma fbdev is convient to add a legency fbdev. And still many Android
devices use fbdev now and at least double buffer is needed for these
Android devices, so that a buffer flip can be operated. It will need some
time for Android device vendors to abondon legency fbdev. So multi buffer
for fbdev is needed.

NOT FOR UPSTREAM:
This is needed only because of the mali driver.
Per Daniel Stone: "... fbdev is deprecated: no new drivers, and no new
features." https://www.spinics.net/lists/dri-devel/msg132518.html

Signed-off-by: Xinliang Liu <xinliang.liu@linaro.org>
[jstultz: Added context in commit message]
Signed-off-by: John Stultz <john.stultz@linaro.org>