drivers/cpufreq: implement init_cpu_capacity_default()
To get default values for CPUs capacity we profile a simple (bogus)
integer benchmark on such CPUs; then we normalize results to 1024
(highest capacity in the system).

Architectures that want this during boot have to register a cpufreq
driver callback and call this function from there (as we require
cpufreq to be up and running).

Signed-off-by: Juri Lelli <juri.lelli@arm.com>