sched: Update max cpu capacity in case of max frequency constraints
Energy-aware wakeup balancing uses cpu capacity awareness and needs to
know the system-wide maximum cpu capacity.

Patch "sched: Store system-wide maximum cpu capacity in root domain"
find the system-wide maximum cpu capacity during scheduler domain
hierarchy setup. This is sufficient as long as maximum frequency
invariance is not enabled.

If it is enabled, the system-wide maximum cpu capacity can change
between scheduler domain hierarchy setups.

The cpu capacity is changed in update_cpu_capacity() which is called in
load balance on the lowest scheduler domain hierarchy level. To be able
to know if a change in cpu capacity for a certain cpu also has an effect
on the system-wide maximum cpu capacity a iteration over all cpus would
be necessary. This would be way too costly. That's why this patch
follows a different approach.

The cpu index I is stored with the system-wide maximum cpu capacity
belonging to this cpu. If the system-wide maximum cpu capacity is
greater than the cpu capacity and the cpu has index I, the system-wide
maximum cpu capacity is set to the cpu capacity. This is also done in
case the system-wide maximum cpu capacity is lesser then the cpu
capacity.

This is the fastest way to update the system-wide maximum cpu capacity
from inside the CFS scheduler code.

But this approach is racy:

 E.g. Two cpus with a higher cpu capacity than the system-wide maximum
      cpu capacity could try to update it at the same time.

and can introduce temporally errors:

 E.g. The capacity of the former cpu which is the one specifying the
      system-wide maximum cpu capacity is lowered and so the
      system-wide maximum cpu capacity is lowered as well even though
      another cpu has now a higher cpu capacity and this cpu will call
      update_cpu_capacity() later).

So far no real issues have been seen and all other approaches (e.g.
propagating the max cpu capacity up the scheduler domain hierarchy seem
to be either too slow or too costly.

There is also an outstanding question:

Should the cpu capacity of a cpu going idle be set to a very small
value?

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>