ACPI / processor: avoid STARTING/DYING actions in a more logical way
As part of the hotplug cleanup, the CPU_STARTING/DYING actions are going
away soon.  This driver needlessly uses those two macro, and so this patch
replaces that code with something more sensible.

Commit 8da8373447d6a57a5a9f55233d35beb15d92d0d2  ("ACPI / processor: Fix
STARTING/DYING action in acpi_cpu_soft_notify()") added checks for those
two actions, because the notification callback can sleep, causing a hung
CPU. This patch instead checks for the ONLINE/DEAD actions, which are the
ones that are handled by the driver in the first place.

Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Cc: Len Brown <lenb@kernel.org>
Cc: linux-acpi@vger.kernel.org
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Richard Cochran <rcochran@linutronix.de>