sched/cpufreq_sched: extend cpufreq_sched API to multiple sched classes
Currently cpufreq_sched only effectively supports a single sched class,
since each capacity request overwrites the previous one.

Extend cpufreq_sched so that the RT and DEADLINE classes may also request
a specific CPU capacity. In addition, the DEADLINE class may request a
minimum capacity level as this is necessary to ensure the semantics of
the DEADLINE class are preserved.

The capacity requests from each scheduling class are summed and that sum
is requested from cpufreq. If there is a DEADLINE minimum capacity request
and that is greater than the sum, it is used instead.

Signed-off-by: Steve Muckle <steve.muckle@linaro.org>