sched/cpufreq_sched: use kthread_bind_mask to set kthread affinity
Use kthread_bind_mask instead of set_cpus_allowed_ptr so that
userspace may not change the affinity of the kthread.

The name of the kthread is also changed since it does not currently
fit within TASK_COMM_LEN, and also does not uniquely identify itself
in the event that there are multiple cpufreq_sched policies running.
The new name includes the number of the first CPU in the policy at
the end of the thread name.

Signed-off-by: Steve Muckle <steve.muckle@linaro.org>