Documentation: arm: define DT cpu capacity bindings
ARM systems may be configured to have cpus with different power/performance
characteristics within the same chip. In this case, additional information
has to be made available to the kernel (the scheduler in particular) for
it to be aware of such differences and take decisions accordingly.

Therefore, this patch aims at standardizing cpu capacities device tree
bindings for ARM platforms. Bindings define cpu capacity parameter, to allow
operating systems to retrieve such information from the device tree and
initialize related kernel structures, paving the way for common code in the
kernel to deal with heterogeneity and removing the need for static data in
current and previous kernel versions (for arch/arm).

Signed-off-by: Juri Lelli <juri.lelli@arm.com>