sched: remove call of sched_avg_update from sched_rt_avg_update
rt_avg is only used to scale the available CPU's capacity for CFS tasks.
As the update of this scaling is done during periodic load balance, we only
have to ensure that sched_avg_update has been called before any periodic load
balancing. This requirement is already fulfilled by __update_cpu_load so the
call in sched_rt_avg_update, which is part of the hotpath, is useless.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>