fix/sched: Always set eenv->cap_idx in find_new_capacity()
Break out of the for loop iterating over the capacity state vector in case
the capacity value is greater or equal the maximum group utilization and
not just return the related capacity state index.
In this case energy_env.cap_idx is also set which is used in
group_norm_util() shortly after the call the find_new_capacity() in
sched_group_energy().

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>