sched: rt scheduler sets capacity requirement
RT task don't provide any running constraints like deadline ones except their
running priority. The only current usable input to estimate the capacity
needed by RT tasks is the rt_avg metric. We use it to estimate the CPU
capacity needed for RT scheduler class.

In order to monitor the evolution or RT tasks load, we must peridoically check
it during the tick.

The, we use the estimated capacity of the last activity to estimate the next
one which can not be that accurate but is a good starting point without any
impact on the wake up path of RT task.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Steve Muckle <smuckle@linaro.org>