linaro/configs: add preempt-rt.config
For the RT patchset, add a preempt-rt fragment to the "mainline"
fragment directory to make it easier to build automatically without
having to go into menuconfig and select the Fully Preemptible Kernel
(RT) to be the Preemption Model and to turn of CPU Frequency scaling. It
isn't mandatory to turn off CPU_FREQ. However, you get a more
predictable latency range.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>
Acked-by: Kevin Hilman <khilman@linaro.org>