fault-injection: cpu: Convert to hotplug state machine
Install the callbacks via the state machine.
This is just temporary. The longterm plan is to create a debugfs interface
where we can put the CPUs up and down from any given point to any given
point. So even if the conversations is not the prettiest - it is just
temporary.

Cc: Akinobu Mita <akinobu.mita@gmail.com>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>