x86: mce: Convert to hotplug state machine
Install the callbacks via the state machine and let the core invoke
the callbacks on the already online CPUs.

mcheck_cpu_init() has been removed from identify_cpu() and pulled in as the
first CPU_STARTING callback. Right now it is invoked a few functions before the
first cpu_notifier is invoked. Except for boot time invocation from the
boot-CPU. Earlier it happened from identify_cpu() now it happens at
device_initcall_sync() time.
One change was required, that is the allocation of memory in
__mcheck_cpu_mce_banks_init(). Initially it happned via identify_cpu() on
boot_cpu where we had interrupts enabled and it happned only once. Now this
callback runs with interrupts disabled and allocations with GFP_KERNEL are no
longer safe. Therefore the allocation of the memory is moved to
mcheck_init_device(). This memory should not be required earlier because MCE
traps & timers are setup in the starting callback.
mce_gen_pool_init() and subsys_system_register() seem not to have a counterpart.
DOWN_PREPARE reverts most of mcheck_cpu_init() doing and DOWN_FAILED remarms it
again. I tried to put the counter part of mcheck_cpu_init() into
mce_cpu_down_dying().

This leaves us with ONLINE and DEAD. From what it looks like DEAD should be
become PREPARE_DOWN since it reverts the ONLINE callbacks. But since it also
invokes ONLINE and DEAD callbacks for the AMD driver I leave it as it (even if
looks that the AMD driver should handle ONLINE and PRE_DOWN as well).

Cc: Tony Luck <tony.luck@intel.com>
Cc: Borislav Petkov <bp@alien8.de>
Cc: linux-edac@vger.kernel.org
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>